SBERDOMA-299 Moved makeClientWithTicket and makeClientWithProperty from tests to utils
