+++ apps/condo/domains/ticket/schema/TicketFile.test.js
@@ -1,17 +1,16 @@
 /**
  * Generated by `createschema ticket.TicketFile 'organization:Text;file?:File;ticket?:Relationship:Ticket:SET_NULL;'`
  */
-const { makeClientWithProperty } = require('@condo/domains/property/schema/Property.test')
+
 const { makeClient, UUID_RE, DATETIME_RE } = require('@core/keystone/test.utils')
 const { 
     TicketFile, 
     createTestTicketFile, 
-    updateTestTicketFile, 
+    updateTestTicketFile,
     makeClientWithTicket,
 } = require('@condo/domains/ticket/utils/testSchema')
 
-
-
+const { makeClientWithProperty } = require('@condo/domains/property/utils/testSchema')
 
 describe('TicketFile', () => {
