+++ apps/condo/domains/ticket/schema/TicketStatus.test.js
@@ -4,7 +4,7 @@
 const faker = require('faker')
 
 const { makeLoggedInAdminClient, makeClient, UUID_RE, DATETIME_RE } = require('@core/keystone/test.utils')
-const { makeClientWithProperty } = require('@condo/domains/property/schema/Property.test')
+const { makeClientWithProperty } = require('@condo/domains/property/utils/testSchema')
 const { TicketStatus, createTestTicketStatus, updateTestTicketStatus } = require('@condo/domains/ticket/utils/testSchema')
 
 describe('TicketStatus', () => {
